+++ libft/ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: phan <phan@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/18 18:25:21 by phan              #+#    #+#             */
-/*   Updated: 2023/03/18 18:59:18 by phan             ###   ########.fr       */
+/*   Updated: 2023/03/18 20:15:32 by phan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *b, int c, unsigned int len)
+#include "libft.h"
+
+void	*ft_memset(void *b, int c, size_t len)
 {
 	unsigned int	i;
 	unsigned char	*_b;
